// apps/mobile/src/api/generated/models/workOrderDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * APU Automotive Service Centre (APU-ASC) API
 * Automotive Service Management RESTful API Specifications
 * OpenAPI spec version: 1.0.0
 */

export interface WorkOrderDto {
	id?: string;
	appointmentId?: string;
	customerId?: string;
	vehicleId?: string;
	serviceId?: string;
	technicianId?: string;
	status?: string;
	intakeNotes?: string;
	diagnosticNotes?: string;
	openedAt?: string;
	startedAt?: string;
	completedAt?: string;
	createdAt?: string;
	updatedAt?: string;
}